Ref #66

- **Remove array Codecs, they use reflect.TypeFor, a go 1.22 feature.**
- **Create API methods for each of the uint sizes.**
- **Create methods for each int size, removing use of reflect.TypeFor.**
- **Remove use of slices.Concat, the last go 1.22 feature in use.**

+++ example_struct_test.go
@@ -30,7 +30,7 @@ func (s SimpleStruct) String() string {
 // Codecs used in these examples.
 // All of these are safe for concurrent access.
 var (
-	anIntCodec      = lexy.Int[int16]()
+	anIntCodec      = lexy.Int16[int16]()
 	aFloatCodec     = lexy.Float32[float32]()
 	stringsCodec    = lexy.Terminate(lexy.SliceOf[[]string](lexy.String[string]()))
 	negStringsCodec = lexy.Negate(stringsCodec)
